Refactor: car.js, race.js -> Car.js, Race.js 파일명 변경
